requirements.txt
numpy==1.18.1
opencv_python==4.2.0.32
keras==2.4.3
keras_applications==1.0.8
Pillow==8.1.0
